mcp_servers/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-servers"
version = "0.1.0"
description = "Modular MCP servers for enterprise HR systems"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "scadam" }
]
license = { text = "MIT" }
dependencies = [
  "mcp>=0.1.0",
  "httpx>=0.27.0",
  "PyJWT>=2.9.0",
  "jwcrypto>=1.5.6",
  "python-dotenv>=1.0.1",
  "pydantic>=2.6.4",
  "structlog>=24.1.0",
  "tenacity>=8.3.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-PyJWT",
  "types-requests"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
